% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard_kweather.R
\name{kweather}
\alias{kweather}
\title{'Shinydashboard' for 'K-WEATHER'}
\usage{
kweather(max.filesize = 2048)
}
\arguments{
\item{max.filesize}{Maximum file size to upload (MB), Default: 2048 (2 GB)}
}
\value{
RStudio addin- 'Shinydashboard' for 'K-WEATHER'
}
\description{
RStudio addin- 'Shinydashboard' for 'K-WEATHER'
}
\details{
RStudio addin- 'Shinydashboard' for 'K-WEATHER'
}
\examples{
if(interactive()){
kweather()
 }
}
\seealso{
\code{\link[data.table]{fread}}
 \code{\link[zip]{unzip}}
 \code{\link[shinyBS]{bsModal}}
 \code{\link[shinyWidgets]{downloadBttn}}
 \code{\link[lubridate]{hour}},\code{\link[lubridate]{day}}
 \code{\link[rmarkdown]{render}},\code{\link[rmarkdown]{html_document}}
}
